Refactor `PlayerTrait`'s volume

+++ playback/src/lib.rs
@@ -564,7 +564,7 @@ impl PlayerTrait for GeneralPlayer {
     async fn add_and_play(&mut self, current_track: &Track) {
         self.get_player_mut().add_and_play(current_track).await;
     }
-    fn volume(&self) -> i32 {
+    fn volume(&self) -> u16 {
         self.get_player().volume()
     }
     fn volume_up(&mut self) {
@@ -573,7 +573,7 @@ impl PlayerTrait for GeneralPlayer {
     fn volume_down(&mut self) {
         self.get_player_mut().volume_down();
     }
-    fn set_volume(&mut self, volume: i32) {
+    fn set_volume(&mut self, volume: u16) {
         self.get_player_mut().set_volume(volume);
     }
     fn pause(&mut self) {
@@ -675,10 +675,10 @@ impl From<PlayerProgress> for crate::player::PlayerTime {
 #[async_trait]
 pub trait PlayerTrait {
     async fn add_and_play(&mut self, current_track: &Track);
-    fn volume(&self) -> i32;
+    fn volume(&self) -> u16;
     fn volume_up(&mut self);
     fn volume_down(&mut self);
-    fn set_volume(&mut self, volume: i32);
+    fn set_volume(&mut self, volume: u16);
     fn pause(&mut self);
     fn resume(&mut self);
     fn is_paused(&self) -> bool;
